pyproject.toml
[project]
name = "terrapyn"
version = "0.1.8"
description = "Toolkit to manipulate Earth observations and models."
authors = [{name = "Colin Hill", email = "colinalastairhill@gmail.com"}]
readme = "README.md"
license = {file = "LICENSE.md"}
requires-python = ">=3.10"
classifiers = [
	"Development Status :: 5 - Production/Stable",
	"Intended Audience :: Developers",
	"License :: OSI Approved :: BSD License",
	"Programming Language :: Python :: 3.10",
	"Programming Language :: Python :: 3.11",
	"Programming Language :: Python :: 3.12",
	"Programming Language :: Python :: 3.13",
]
dependencies = [
	"xarray>=2025.6.1",
	"pandas>=2.3.1",
	"numpy>=2.2.6",
	"scipy>=1.15.3",
	"dask>=2025.7.0",
	"netcdf4>=1.7.2",
	"zarr>=2.18.3",
	"bottleneck>=1.5.0",
	"geopandas>=1.1.1",
	"matplotlib>=3.10.5",
	"shapely>=2.1.1",
	"ipykernel>=6.30.1",
	"pyarrow>=21.0.0",
	"fiona>=1.10.1",
	"tqdm>=4.67.1",
	"jupyterlab>=4.4.6",
	"rioxarray>=0.19.0",
	"earthengine-api>=1.6.4",
	"geemap>=0.36.1",
	"google-cloud-bigquery>=3.36.0",
	"db-dtypes>=1.4.3",
	"scikit-learn>=1.7.1",
	"seaborn>=0.13.2",
	"odc-geo>=0.4.10",
	"gcsfs>=2025.7.0",
	"bokeh>=3.7.3",
	"openpyxl>=3.1.5",
	"distributed>=2025.7.0",
]

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.hatch.build.targets.wheel]
packages = ["terrapyn"]

[tool.hatch.build.targets.sdist]
exclude = [
  "/.github",
  "/docs",
]

[tool.hatch.version]
path = "pyproject.toml"

[tool.ruff]
line-length = 120
target-version = "py312"
extend-exclude = [
    ".venv",
]

[tool.ruff.lint]
select = [
  "B",    # flake8-bugbear
  "E",    # pycodestyle errors
  "F",    # pyflakes
  "UP",   # pyupgrade
  "D",    # pydocstyle
  "I",    # isort
  "PD",   # pandas-vet
  "SIM",  # flake8-simplify
  "NPY",  # numpy-specific
]
ignore = [
	"D1",
	"D2",
	"D4",
	"TID",
    "D400", # Period at the end of the first line of a docstring
	"PD901", # generic variable name `df` for DataFrames
]

[tool.ruff.format]
indent-style = "tab"

[tool.ruff.lint.pydocstyle]
convention = "google"

[dependency-groups]
dev = [
	"pytest",
	"pytest-cov",
	"isort",
	"black",
	"freezegun",
	"nbstripout",
	"coverage-badge",
	"pre-commit>=4.2.0",
    "ruff>=0.12.9",

]

[tool.pytest.ini_options]
# Ignore earth engine directory 'ee'
addopts = "--doctest-modules --cov=terrapyn --cov-branch --cov-report term-missing  -vv --color=yes --ignore=terrapyn/ee"
